+++ src/Areas/VirtualMachines/Options/BaseVirtualMachinesOptions.cs
@@ -1,0 +1,11 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using AzureMcp.Options;
+
+namespace AzureMcp.Areas.VirtualMachines.Options;
+
+public class BaseVirtualMachinesOptions : SubscriptionOptions
+{
+    public string? ResourceGroup { get; set; }
+}